pyproject.toml
[project]
name = "unfat"
version = "0.0.4"
description = "Extract datasets from models and train slimmer LoRAs on them"
authors = [
    {name = "@reissbaker"}
]
readme = "README.md"
license = "MIT"
keywords = ["finetune", "LoRA", "llm"]
requires-python = ">=3.13,<4.0"
dependencies = [
    "tqdm (>=4.67.1,<5.0.0)",
    "datasets (>=3.2.0,<4.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "pylint (>=3.3.4,<4.0.0)",
    "mypy (>=1.14.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "together (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
